Feat/redemption fee

+++ contracts/libraries/uni-v2/UniswapV2Library.sol
@@ -5,9 +5,9 @@ library MinimalUniswapV2Library {
     error InvalidToken();
 
     // 0.3%
-    uint public constant FEE = 997;
+    uint256 public constant FEE = 997;
     // 0%
-    uint public constant NO_FEE = 1000;
+    uint256 public constant NO_FEE = 1000;
 
     // returns sorted token addresses, used to handle return values from pairs sorted in this order
     function sortTokensUnsafeWithAmount(
@@ -73,35 +73,35 @@ library MinimalUniswapV2Library {
     // given an input amount of an asset and pair reserves, returns the maximum output amount of the other asset
     // WARNING: won't apply fee since we don't take fee from user right now
     function getAmountOut(
-        uint amountIn,
-        uint reserveIn,
-        uint reserveOut
-    ) internal pure returns (uint amountOut) {
+        uint256 amountIn,
+        uint256 reserveIn,
+        uint256 reserveOut
+    ) internal pure returns (uint256 amountOut) {
         require(amountIn > 0, "UniswapV2Library: INSUFFICIENT_INPUT_AMOUNT");
         require(
             reserveIn > 0 && reserveOut > 0,
             "UniswapV2Library: INSUFFICIENT_LIQUIDITY"
         );
-        uint amountInWithFee = amountIn * NO_FEE;
-        uint numerator = amountInWithFee * reserveOut;
-        uint denominator = reserveIn * 1000;
+        uint256 amountInWithFee = amountIn * NO_FEE;
+        uint256 numerator = amountInWithFee * reserveOut;
+        uint256 denominator = reserveIn * 1000;
         amountOut = numerator / denominator;
     }
 
     // given an output amount of an asset and pair reserves, returns a required input amount of the other asset
     // WARNING: won't apply fee since we don't take fee from user right now
     function getAmountIn(
-        uint amountOut,
-        uint reserveIn,
-        uint reserveOut
-    ) internal pure returns (uint amountIn) {
+        uint256 amountOut,
+        uint256 reserveIn,
+        uint256 reserveOut
+    ) internal pure returns (uint256 amountIn) {
         require(amountOut > 0, "UniswapV2Library: INSUFFICIENT_OUTPUT_AMOUNT");
         require(
             reserveIn > 0 && reserveOut > 0,
             "UniswapV2Library: INSUFFICIENT_LIQUIDITY"
         );
-        uint numerator = reserveIn * amountOut * 1000;
-        uint denominator = reserveOut * NO_FEE;
+        uint256 numerator = reserveIn * amountOut * 1000;
+        uint256 denominator = reserveOut * NO_FEE;
         amountIn = (numerator / denominator);
     }
 }
